Update Fix_abrade_MPI.h
Began adding the wall-particle abrasion. This required extract methods to be added to fix_wall/gran/region

+++ Fix_abrade_MPI.h
@@ -81,7 +81,7 @@ private:
 
   class NeighList *list;
   void areas_and_normals();
-  void displacement_of_atom(int, int, double[3], double[3]);
+  void displacement_of_atom(int, double, double[3], double[3]);
 
  protected:
   int me, nprocs;
